Move python files in top level directory to a new directory. (#108)
 This structure helps us in python 3 land to have a more conventional directory structure for unit testing.
