+++ tests/datastore_emulator/datastore_cache_api_test.py
@@ -1,7 +1,7 @@
 import time
 import unittest
 
-from datastore_cache_api import DatastoreCacheApi
+from bond_app.datastore_cache_api import DatastoreCacheApi
 from unit import cache_api_test
 import datastore_emulator_utils
