+++ tests/unit/bond_test.py
@@ -7,13 +7,13 @@ import jwt
 from mock import MagicMock
 from werkzeug import exceptions
 
-from authentication import UserInfo
-from bond import Bond, FenceKeys
-from fence_api import FenceApi
-from fence_token_vending import FenceTokenVendingMachine
-from oauth_adapter import OauthAdapter
-from sam_api import SamApi
-from sam_api import SamKeys
+from bond_app.authentication import UserInfo
+from bond_app.bond import Bond, FenceKeys
+from bond_app.fence_api import FenceApi
+from bond_app.fence_token_vending import FenceTokenVendingMachine
+from bond_app.oauth_adapter import OauthAdapter
+from bond_app.sam_api import SamApi
+from bond_app.sam_api import SamKeys
 from fake_token_store import  FakeTokenStore
 from fake_cache_api import FakeCacheApi
 from fake_fence_token_storage import FakeFenceTokenStorage
